+++ jvm-blesser/src/test/scala/santa/EvalConformanceTest.scala
@@ -73,7 +73,7 @@ class EvalConformanceTest extends munit.FunSuite {
           val inputs = ec.downField("inputs").values.getOrElse(Vector.empty).toVector.map { inp =>
             val ext = inp.hcursor.downField("extension")
             ext.keys.getOrElse(Iterable.empty).iterator.map { k =>
-              k.toByte -> EvalCore.decodeInputConstant(
+              k.toInt.toByte -> EvalCore.decodeInputConstant(
                 ext.downField(k).focus.getOrElse(sys.error(s"v3 entry '$name': missing extension value for key $k")))
             }.toMap
           }
